this and that bugfixes etc

+++ WFC/vec2i_util.h
@@ -2,7 +2,7 @@
 #include <SFML/Graphics.hpp>
 #include "Entity.h"
 static int sumSq(sf::Vector2i val) {
-	return pow(val.x, 2) + pow(val.y, 2);
+	return static_cast<int>(pow(val.x, 2) + pow(val.y, 2));
 }
 static int distSq(const Entity* a, const Entity* b) {
 	return sumSq(a->getPos() - b->getPos());
